# Tracking performance degradation on simple contract call