--- a/Cargo.toml
+++ b/Cargo.toml
@@ -67,5 +67,9 @@ harness = false
 name = "conversion"
 harness = false
 
+[[bench]]
+name = "simple"
+harness = false
+
 [profile.bench]
 debug = false