--- a/Cargo.toml
+++ b/Cargo.toml
@@ -44,8 +44,9 @@ hex-literal = "0.4.1"
 criterion = {version="0.3.6", features=["html_reports"] }
 
 [features]
-default = []
+default = ["with-serde"]
 redis=["dep:redis"]
+with-serde = []
 
 [[bench]]
 name = "general"