--- a/src/fork_db.rs
+++ b/src/fork_db.rs
@@ -13,6 +13,7 @@ use revm::primitives::{
 };
 use revm::{Database, DatabaseCommit};
 use std::env;
+use std::hash::Hash;
 use tracing::{debug, info, trace};
 
 #[derive(Debug, Default)]
@@ -174,7 +175,9 @@ impl<T: ProviderCache> ForkDB<T> {
         let mut changed = false;
         if let Some(code) = &account.code {
             if !code.is_empty() {
-                account.code_hash = keccak256(code.bytecode());
+                if account.code_hash == KECCAK_EMPTY {
+                    account.code_hash = code.hash_slow();
+                }
                 self.contracts
                     .entry(account.code_hash)
                     .or_insert_with(|| code.clone());
@@ -317,6 +320,7 @@ impl<T: ProviderCache> DatabaseCommit for ForkDB<T> {
             if !account.is_touched() {
                 continue;
             }
+
             if account.is_selfdestructed() {
                 let db_account = self.accounts.entry(address).or_default();
                 db_account.storage.clear();