… tests/contracts/TetherToken.hex