```
❯ cargo bench call_call_get_balance
Benchmarking call_call_get_balance: Collecting 100 samples in estimated 10.266 s (172k iterations)
call_call_get_balance   time:   [59.127 µs 59.696 µs 60.315 µs]
                        change: [+0.9024% +2.1704% +3.5008%] (p = 0.00 < 0.05)
                        Change within noise threshold.
```

compared to v1 which takes less than half of the time:

```
call_call_get_balance   time:   [21.068 µs 21.286 µs 21.525 µs]
```