@cassc 2024-07-24
Tracked down to https://github.com/sbip-sg/tevm/blob/abc7d3e07975197960c49125b47d5c191deea9ad/src/fork_db.rs#L328, looks like `insert_contract` is executed unnecessarily too often in v2.

For v2
![flamegraph](https://github.com/user-attachments/assets/4406e812-57e2-45d0-b536-2e1445e3073c)

For v1
![flamegraph](https://github.com/user-attachments/assets/8ddcf364-6c48-4a88-b7ee-83ecefb4e18b)